User/relationship associations

+++ db/seeds.rb
@@ -5,7 +5,7 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-User.delete_all
+User.delete_all if User.any?
 
 User.create!(name: "Example User",
              email: "example@railstutorial.org",
